+++ corehq/apps/userreports/tests/test_expressions.py
@@ -6,12 +6,10 @@ from decimal import Decimal
 from django.test import SimpleTestCase, TestCase, override_settings
 from fakecouch import FakeCouchDb
 from simpleeval import InvalidExpression
-import mock
 from casexml.apps.case.const import CASE_INDEX_EXTENSION
 from casexml.apps.case.mock import CaseStructure, CaseFactory, CaseIndex
 from casexml.apps.case.models import CommCareCase
 from casexml.apps.case.tests.util import delete_all_cases, delete_all_xforms
-from corehq.apps.es.fake.forms_fake import FormESFake
 from corehq.apps.userreports.exceptions import BadSpecError
 from corehq.apps.userreports.expressions.factory import ExpressionFactory
 from corehq.apps.userreports.expressions.specs import (
@@ -1053,116 +1051,6 @@ class TestFormsExpressionSpec(TestCase):
 
 @override_settings(TESTS_SHOULD_USE_SQL_BACKEND=True)
 class TestFormsExpressionSpecSQL(TestCase):
-    pass
-
-
-@override_settings(TESTS_SHOULD_USE_SQL_BACKEND=False)
-@mock.patch('corehq.apps.userreports.expressions.specs.FormES', FormESFake)
-class TestFormsExpressionSpecWithFilter(TestCase):
-
-    @classmethod
-    def setUpClass(cls):
-        super(TestFormsExpressionSpecWithFilter, cls).setUpClass()
-        cls.domain = uuid.uuid4().hex
-        factory = CaseFactory(domain=cls.domain)
-        [cls.case] = factory.create_or_update_case(CaseStructure(attrs={'create': True}))
-        cls.forms = [f.to_json() for f in FormAccessors(cls.domain).get_forms(cls.case.xform_ids)]
-        for form in cls.forms:
-            FormESFake.save_doc(form)
-        #  redundant case to create extra forms that shouldn't be in the results for cls.case
-        [cls.case_b] = factory.create_or_update_case(CaseStructure(attrs={'create': True}))
-        cls.forms_b = [f.to_json() for f in FormAccessors(cls.domain).get_forms(cls.case_b.xform_ids)]
-        for form in cls.forms_b:
-            FormESFake.save_doc(form)
-
-    @classmethod
-    def tearDownClass(cls):
-        FormESFake.reset_docs()
-        delete_all_xforms()
-        delete_all_cases()
-        super(TestFormsExpressionSpecWithFilter, cls).tearDownClass()
-
-    def _make_expression(self, from_statement, to_statement, xmlns=None):
-        spec = {
-            "type": "get_case_forms_in_date",
-            "case_id_expression": {
-                "type": "property_name",
-                "property_name": "_id"
-            },
-            "from_date_expression": {
-                "type": "add_months",
-                "datatype": "datetime",
-                "date_expression": {
-                    "type": "property_name",
-                    "property_name": "server_modified_on"
-                },
-                "months_expression": {
-                    "type": "evaluator",
-                    "statement": from_statement,
-                    "context_variables": {
-                        "iteration": {
-                            "type": "base_iteration_number"
-                        }
-                    }
-                },
-            },
-            "to_date_expression": {
-                "type": "add_months",
-                "datatype": "datetime",
-                "date_expression": {
-                    "type": "property_name",
-                    "property_name": "server_modified_on"
-                },
-                "months_expression": {
-                    "type": "evaluator",
-                    "statement": to_statement,
-                    "context_variables": {
-                        "iteration": {
-                            "type": "base_iteration_number"
-                        }
-                    }
-                },
-            },
-        }
-
-        if xmlns:
-            spec['xmlns'] = [xmlns]
-
-        return ExpressionFactory.from_spec(spec)
-
-    def test_from_inside_date_range(self):
-        expression = self._make_expression('iteration - 1', 'iteration + 1')
-        context = EvaluationContext({"domain": self.domain}, 0)
-        forms = expression(self.case.to_json(), context)
-
-        self.assertEqual(len(forms), 1)
-        self.assertEqual(forms, self.forms)
-
-    def test_from_outside_date_range(self):
-        expression = self._make_expression('iteration - 2', 'iteration - 1')
-        context = EvaluationContext({"domain": self.domain}, 0)
-        forms = expression(self.case.to_json(), context)
-
-        self.assertEqual(forms, [])
-
-    def test_from_correct_xmlns(self):
-        expression = self._make_expression('iteration - 1', 'iteration + 1', 'http://commcarehq.org/case')
-        context = EvaluationContext({"domain": self.domain}, 0)
-        forms = expression(self.case.to_json(), context)
-
-        self.assertEqual(len(forms), 1)
-        self.assertEqual(forms, self.forms)
-
-    def test_from_incorrect_xmlns(self):
-        expression = self._make_expression('iteration - 1', 'iteration + 1', 'silly-xmlns')
-        context = EvaluationContext({"domain": self.domain}, 0)
-        forms = expression(self.case.to_json(), context)
-
-        self.assertEqual(forms, [])
-
-
-@override_settings(TESTS_SHOULD_USE_SQL_BACKEND=True)
-class TestFormsExpressionSpecWithFilterSQL(TestFormsExpressionSpecWithFilter):
     pass
